Reverse Handler Works
This commit doesn’t really fix the handler, since it was already
working. Outputs were never being recorded by ROSS in optimistic debug
mode, which is why it never appeared to be working. Will keep the misc
changes in the commit, thought the ROSS source code change that allowed
the output to be recorded and outputted (showing that the reverse
handler works) was the addition of the call ‘tw_stats(me);’ on line 498
of tw-sched.c

+++ myModel3.h
@@ -60,15 +60,15 @@ typedef struct {
 } state;
 
 typedef struct {
-    long int health;
-    long int health_lim;
-    long int resources;
-    long int offense;
-    long int size;
-    long int at_war_with;
-    long int times_won;
-    long int times_defeated;
-    long int wars_started;
+    long long int health;
+    long long int health_lim;
+    long long int resources;
+    long long int offense;
+    long long int size;
+    long long int at_war_with;
+    long long int times_won;
+    long long int times_defeated;
+    long long int wars_started;
 } final_stats;
 
 /*
